reorganize utilities funcs
